refactor: enhance embedding dimension handling, improve neo4j schema index queries, and align imports with preferred aliases
