+++ AGENTS.md
@@ -6,4 +6,5 @@
 -this is an early stage project so breaking changes are expected (but do alert me at the end of your message of the breaking changes )
 - prefer import aliases
 - when writing cypher queries start the string with ```/*cypher*/``` so the neo4j extension can activate
+- when writing cypher queries , prefer session.executeWrite() & session.executeRead() over session.run()
 - when tasked with frontend/UI work, use the frontend design skill
