+++ docs/schema.md
@@ -1,0 +1,4 @@
+# Schema
+## Nodes
+### AstNode
+-
